fix: widen player_id and selected_id from int to uint64_t
Eliminates silent truncation when entity IDs exceed INT32_MAX.
Sentinel value changed from -1 to 0 (ID 0 is never assigned).

+++ src/dui_mock.cpp
@@ -30,8 +30,8 @@ World MakeMockWorld() {
          .SetType(static_cast<uint8_t>(i % 3))
          .SetLabel("#%d", 1000 + i);
 
-    w.player_id     = static_cast<int>(w.entities[0].id);
-    w.selected_id   = w.player_id;
+    w.player_id   = w.entities[0].id;
+    w.selected_id = w.player_id;
     w.selected_ids.push_back(w.entities[0].id);
 
     const uint32_t wall_col  = RGBA(90,  70,  70,  200);
